app/requirements.txt
fastapi
uvicorn[standard]
Pillow
python-multipart
python-dotenv
openai
pydantic
pydicom